move configuration to config package
